+++ modules/dsfhub-gcp-cloud-account/outputs.tf
@@ -1,0 +1,4 @@
+output "this" {
+  description = "GCP cloud account asset."
+  value       = dsfhub_cloud_account.this
+}