+++ modules/dsfhub-gcp-pubsub/outputs.tf
@@ -1,0 +1,4 @@
+output "this" {
+  description = "GCP PUBSUB asset."
+  value       = dsfhub_log_aggregator.this
+}