+++ modules/google-pubsub-topic/outputs.tf
@@ -1,0 +1,4 @@
+output "this" {
+  description = "Google PubSub Topic."
+  value       = google_pubsub_topic.this
+}